+++ AGGR_OP/funcaggregoperator.h
@@ -21,7 +21,7 @@ private:
 	// == METHODS ==
 public:
 	FuncAggregOperator(AggregOperatorFunc::AggrOpFunc t_func);
-	int GetWorthlyValue(const QList<int> &t_list);
+	int GetWorthlyValue(const QList<long double> &t_list);
 
 protected:
 	void ResetValues();
