+++ AGGR_OP/harmmeanaggregoperator.h
@@ -19,7 +19,7 @@ private:
 	// == METHODS ==
 public:
 	HarmMeanAggregOperator();
-	int GetWorthlyValue(const QList<int> &t_list);
+	int GetWorthlyValue(const QList<long double> &t_list);
 
 protected:
 	void ResetValues();
