+++ AGGR_OP/meanaggregoperator.h
@@ -20,7 +20,7 @@ private:
 	// == METHODS ==
 public:
 	MeanAggregOperator(const double &t_power);
-	int GetWorthlyValue(const QList<int> &t_list);
+	int GetWorthlyValue(const QList<long double> &t_list);
 
 protected:
 	void ResetValues();
